+++ BancoDeHoras.Api/Controllers/MarcaPontoController.cs
@@ -2,6 +2,7 @@
 using BancoDeHoras.Application.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.ComponentModel.DataAnnotations;
 
 namespace BancoDeHoras.Api.Controllers
 {
@@ -23,21 +24,35 @@ namespace BancoDeHoras.Api.Controllers
         }
 
         [HttpPost]
-        public void Post(MarcaPontoDTO entity)
+        public ActionResult Post(MarcaPontoDTO entity)
         {
+            if (entity.Id == 0 || entity.Id == null)
+                return BadRequest("Insira um ID para continuar a operação");
+
             _service.Post(entity);
+            return Ok("Dados inseridos com sucesso!!!");
         }
 
         [HttpPut]
-        public void Put(MarcaPontoDTO entity)
+        public ActionResult Put([Required][FromQuery]MarcaPontoDTO entity)
         {
+            if (entity.Id == 0 || entity.Id == null)
+                return BadRequest("Insira um ID para continuar a operação");
+
             _service.Put(entity);
+            return Ok("Dados atualizados com sucesso!!!");
+
         }
 
         [HttpDelete]
-        public void Delete(MarcaPontoDTO entity)
+        public ActionResult Delete(MarcaPontoDTO entity)
         {
+            if (entity.Id == 0 || entity.Id == null)
+                return BadRequest("Insira um ID para continuar a operação");
+
             _service.Delete(entity);
+            return Ok("Dado deletado com sucesso!!!");
+
         }
     }
 }
